Enable support for arbitrary matching values in spec.
Instead of only allowing dicts as values in a specification, now all
types are accepted.

dicts are considered subdocuments and are recursed into. All other
types must be tested equal to the corresponding document node. If the
test is successfule the document node is preserved in the instance. If
the document node is empty and autoconstruct is true, then the spec node
is placed at the instance.