Define a config file that includes spec and other settings
This config file must located in the home directory of user and named
as `.apimas`. This file must follows `YAML` syntax. This file is
required and it must define the root url of your application, e.g.
`http://localhost:8000` and the APIMAS specification in order to
construct clients and cli to interact with the API.