Access private 'self._context' instead of property
The access of this property triggered the access of another property
named `root`, which is django cached_property. A nested serializer is
initialized firstly without being bound to any parent serializer.

Therefore, the cached_property property would return `None`, as there
is not any parent serialzier. However, the time when this nested
serializer is bound to a parent serializer, the execution of this
property is not triggered (because it is cached), therefore, we are
unable to access the parent serializer.

The bug that introduced this problem was that when we performed a
partial update on a resource with nested serializers, there were some
fields (in particular nested) that could not retrieve all the context
of request which is specified on the root serializer.