Fix permission checks
There was an error on how the permission rules were checked by
`ApimasPermission` class.

Now it checks which states are valid during the course of request and
then it finds the subset of matches that satisfy the valid states.
Given this subset of matches, writable and accesible fields are defined
accordingly (Subset of fields that are satisfied by the subset of
matches).