Add support for pipeline architecture
An action can be described by a collection on which it is performed,
a URL, a set of processors and one handler.

Processors are classified into request and response processors. The
former receive the request context and may read and write data on it,
whilst the latter receive the response context.

The handler receives the request context and produces a response
context. It is the actual hook of the action, which describes the
business logic. For instance, a handler may interact with a db and
create a new record.