ClientHandler: support data fetching for ref fields
This commit adds support for data fetching. Typically, the value of a
ref field returned by the server is a hyperlink which points to the
location of the related field. For example:

```lang=json
{
    "myref": "api/mycollection/myid/"
}
```

However, many times we need to get data from the related fields. For
this reason we need to make a `GET` request to the hypelink indicated by
the response:

```lang=json
{
    "myref": {
        "foo": "value"
    }
}
```

To enable this feature we attach the `fetch_refs` key inside `.meta`.
For instance:

```lang=python
spec = {
    '.meta': {
        'fetch_refs': True
    }
    'api': {
        '.endpoint': {}
        'mycollection': {
            '.collection': {},
            '*': {
                'myref': {
                    '.ref': {
                        'to': 'api/ref'
                    },
                    '.fetch': {
                        'included': [
                            'foo',
                        ],
                        'excluded': [
                            'bar,'
                        ]
                    }
                }
            }
        }
    }
}
```

Note predicate `.fetch` inside `myref` key which indicates which fields
of the response should be included or excluded.