Pass nullable attribute to Serializers
Specifically the String serializer needs to know if None is a valid
value to be serialized. If a text/char field is marked as nullable then
None should be returned as null (or the respective null-value of the
returned seralization format). Otherwise None is not a valid valid and
it cannot be serialized.